Benchmarks: update simdjson to latest version + cleanup (#113)

+++ bench/simdjson-utf8/src/lib.rs
@@ -1,11 +1,11 @@
-#![expect(non_camel_case_types)]
 use std::os::raw::c_char;
+use std::os::raw::c_ulong;
 
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
 #[inline(always)] // only works if simdjson ist compiled with LLVM and cross-language LTO is enabled
 pub fn validate(bytes: &[u8]) -> bool {
-    unsafe { simdjson_validate_utf8(bytes.as_ptr() as *const c_char, bytes.len() as size_t) }
+    unsafe { simdjson_validate_utf8(bytes.as_ptr() as *const c_char, bytes.len() as c_ulong) }
 }
 
 #[cfg(test)]
